api/requirements.txt
flask==3.0.2
Frozen-Flask==0.18
python-dotenv==1.0.1
gunicorn==21.2.0 